refactor: refactor error page
